+++ src/screens/recipient/components/SessionActionArea.tsx
@@ -275,4 +275,4 @@ const createStyles = (colors: typeof Colors) => StyleSheet.create({
   },
 });
 
-export default SessionActionArea;
+export default React.memo(SessionActionArea);